+++ qgreenland/tasks/common/shapefile.py
@@ -1,5 +1,4 @@
 import os
-import zipfile
 
 import luigi
 
@@ -7,22 +6,6 @@ from qgreenland.constants import TaskType
 from qgreenland.util.luigi import LayerTask
 from qgreenland.util.misc import find_single_file_by_ext, temporary_path_dir
 from qgreenland.util.shapefile import reproject_shapefile, subset_shapefile
-
-
-class UnzipShapefile(LayerTask):
-    task_type = TaskType.WIP
-
-    def output(self):
-        return luigi.LocalTarget(f'{self.outdir}/unzip/')
-
-    def run(self):
-        zf_path = find_single_file_by_ext(self.input().path, ext='.zip')
-        zf = zipfile.ZipFile(zf_path)
-
-        with temporary_path_dir(self.output()) as temp_path:
-            for fn in zf.namelist():
-                zf.extract(fn, temp_path)
-            zf.close()
 
 
 class ReprojectShapefile(LayerTask):
@@ -40,6 +23,9 @@ class ReprojectShapefile(LayerTask):
     def run(self):
         shapefile = find_single_file_by_ext(self.input().path, ext='.shp')
         gdf = reproject_shapefile(shapefile)
+
+        # TODO: Dissolve polygon boundaries if required by config file
+
         with temporary_path_dir(self.output()) as temp_path:
             fn = os.path.join(temp_path, self.filename)
             gdf.to_file(fn, driver='ESRI Shapefile')
@@ -53,8 +39,6 @@ class SubsetShapefile(LayerTask):
 
     def run(self):
         shapefile = find_single_file_by_ext(self.input().path, ext='.shp')
-        gdf = subset_shapefile(shapefile, layer_cfg=self.layer_cfg)
-
         with temporary_path_dir(self.output()) as temp_path:
             fn = os.path.join(temp_path, self.filename)
-            gdf.to_file(fn, driver='ESRI Shapefile')
+            subset_shapefile(shapefile, layer_cfg=self.layer_cfg, outfile=fn)